save OrderServiceImpl

+++ src/main/java/kg/mega/natv_final_project/models/dto/responses/response1/DiscountListDto.java
@@ -9,6 +9,6 @@ import lombok.experimental.FieldDefaults;
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DiscountListDto {
-    int fromDayCount;
+    double fromDayCount;
     double discount;
 }
